refactor: search hook
fixes #34
