+++ packages/brokeneck-react/src/components/Groups.js
@@ -1,10 +1,9 @@
-import React, { useEffect, useMemo, useState } from 'react'
+import React from 'react'
 import {
   Box,
   Button,
   CircularProgress,
   IconButton,
-  InputAdornment,
   Link,
   makeStyles,
   Table,
@@ -13,19 +12,18 @@ import {
   TableContainer,
   TableHead,
   TableRow,
-  TextField,
   Typography
 } from '@material-ui/core'
 import { Link as RouterLink, useRouteMatch } from 'react-router-dom'
 import { useQuery } from 'graphql-hooks'
 import startCase from 'lodash.startcase'
-import debounce from 'lodash.debounce'
 
 import useCreateGroupDialog from '../hooks/useCreateGroupDialog'
 import useFields, { TYPE_NAMES } from '../hooks/useFields'
 import useSurrogatePagination from '../hooks/useSurrogatePagination'
 import { LOAD_GROUPS } from '../graphql'
 import useProvider from '../hooks/useProvider'
+import useSearch from '../hooks/useSearch'
 
 import Square from './Square'
 
@@ -45,17 +43,11 @@ const useStyles = makeStyles(theme => ({
 export default function Groups() {
   const match = useRouteMatch()
   const groupFields = useFields('Group')
-  const [search, setSearch] = useState({ immediate: '', debounced: '' })
+  const { search, Search } = useSearch()
   const classes = useStyles()
   const {
     capabilities: { canSearchGroups }
   } = useProvider()
-
-  const debouncedSetSearch = useMemo(() => debounce(setSearch, 500), [
-    setSearch
-  ])
-
-  useEffect(() => () => debouncedSetSearch.cancel(), [debouncedSetSearch])
 
   const {
     pageSize,
@@ -70,15 +62,10 @@ export default function Groups() {
       variables: {
         pageNumber: surrogatePageNumber,
         pageSize,
-        search: search.debounced
+        search
       }
     }
   )
-
-  const handleSearchChange = search => {
-    setSearch(s => ({ ...s, immediate: search }))
-    debouncedSetSearch(s => ({ ...s, debounced: search }))
-  }
 
   const [dialog, openDialog] = useCreateGroupDialog(loadGroups)
 
@@ -94,28 +81,7 @@ export default function Groups() {
           <Button variant="contained" color="primary" onClick={openDialog}>
             Create Group
           </Button>
-          {canSearchGroups && (
-            <TextField
-              label="Search"
-              variant="outlined"
-              size="small"
-              value={search.immediate}
-              onChange={e => handleSearchChange(e.target.value)}
-              InputProps={{
-                endAdornment: search.immediate && (
-                  <InputAdornment position="end">
-                    <IconButton onClick={() => handleSearchChange('')}>
-                      <Typography>
-                        <span role="img" aria-label="clear">
-                          ❌
-                        </span>
-                      </Typography>
-                    </IconButton>
-                  </InputAdornment>
-                )
-              }}
-            />
-          )}
+          {canSearchGroups && Search}
           <div className={classes.right}>
             <IconButton onClick={loadGroups} title="reload groups">
               <Typography>
